pyproject.toml
[project]
name = "error-correction"
version = "0.1.0"
description = "基于 LLM + OCR 的智能错题本生成系统"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "flask==3.1.2",
    "langchain==1.2.3",
    "langchain-core==1.2.7",
    "langchain-deepseek==1.0.1",
    "langgraph==1.0.5",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pdf2image>=1.16.3",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv==1.2.1",
    "requests>=2.31.0",
    "rich==14.2.0",
    "sqlalchemy>=2.0",
    "tqdm>=4.65.0",
    "werkzeug==3.1.5",
]

[dependency-groups]
dev = [
    "pytest==9.0.2",
]
